#architecture/dependency_injection/di_hilt: The dependency injection was implemented using hilt.
