+++ app/src/main/AndroidManifest.xml
@@ -5,14 +5,14 @@
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
-        android:name=".androidframework.ArchApp"
+        android:name=".framework.ArchApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher_arch"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_arch_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.ArchApp">
-        <activity android:name=".presentation.MainActivity">
+        <activity android:name="com.example.luistovar.archapp.framework.presentation.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
