--- app/src/main/java/com/example/luistovar/archapp/androidframework/di/AppModule.kt
@@ -1,11 +1,0 @@
-package com.example.luistovar.archapp.androidframework.di
-
-import org.koin.dsl.module
-
-val appModule = module {
-
-}
-
-
-
-
