--- app/src/main/java/com/example/luistovar/archapp/androidframework/di/NetworkModule.kt
@@ -1,8 +1,0 @@
-package com.example.luistovar.archapp.androidframework.di
-
-import com.example.luistovar.archapp.androidframework.network.NetWorkConfiguration
-import org.koin.dsl.module
-
-val networkModule = module {
-    factory { NetWorkConfiguration.getStarWarsApi() }
-}
