--- app/src/main/java/com/example/luistovar/archapp/androidframework/di/RepositoryModule.kt
@@ -1,19 +1,0 @@
-package com.example.luistovar.archapp.androidframework.di
-
-import com.example.luistovar.archapp.data.datasources.remote.PeopleListSwDataSource
-import com.example.luistovar.archapp.data.datasources.remote.impl.PeopleListSwDataSourceRemoteImpl
-import com.example.luistovar.archapp.data.repositories.PeopleListSwRepository
-import com.example.luistovar.archapp.data.repositories.implementation.PeopleListSwRepositoryImpl
-import org.koin.dsl.module
-
-val localDataSourceModule = module {
-
-}
-
-val remoteDataSourceModule = module {
-    factory<PeopleListSwDataSource> { PeopleListSwDataSourceRemoteImpl(get()) }
-}
-
-val repositoryModule = module {
-    factory<PeopleListSwRepository> { PeopleListSwRepositoryImpl(get()) }
-}
