--- app/src/main/java/com/example/luistovar/archapp/data/datasources/remote/PeopleListSwDataSource.kt
@@ -1,9 +1,0 @@
-package com.example.luistovar.archapp.data.datasources.remote
-
-import com.example.luistovar.archapp.domain.models.PeopleListSw
-import com.example.luistovar.archapp.domain.models.Resource
-
-interface PeopleListSwDataSource {
-
-    suspend fun getPeopleListSwResource(): Resource<PeopleListSw?>
-}
