+++ app/src/main/java/com/example/luistovar/archapp/data/datasources/remote/PeopleListSwRemoteDataSource.kt
@@ -1,0 +1,9 @@
+package com.example.luistovar.archapp.data.datasources.remote
+
+import com.example.luistovar.archapp.domain.models.PeopleListSw
+import com.example.luistovar.archapp.domain.models.Resource
+
+interface PeopleListSwRemoteDataSource {
+
+    suspend fun getPeopleListSwResource(): Resource<PeopleListSw?>
+}
