--- app/src/main/java/com/example/luistovar/archapp/domain/usecases/PeopleListSwUseCase.kt
@@ -1,11 +1,0 @@
-package com.example.luistovar.archapp.domain.usecases
-
-import com.example.luistovar.archapp.domain.models.PeopleListSw
-import com.example.luistovar.archapp.domain.models.Resource
-
-interface PeopleListSwUseCase {
-
-    suspend fun getPeopleListSwResource(): Resource<PeopleListSw?>
-
-
-}
