+++ app/src/main/java/com/example/luistovar/archapp/framework/di/AppModule.kt
@@ -1,0 +1,11 @@
+package com.example.luistovar.archapp.framework.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+
+}
