+++ app/src/main/java/com/example/luistovar/archapp/framework/di/network/NetworkModule.kt
@@ -1,0 +1,60 @@
+package com.example.luistovar.archapp.framework.di.network
+
+import com.example.luistovar.archapp.BuildConfig
+import com.example.luistovar.archapp.framework.network.CommonHeadersInterceptor
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import timber.log.Timber
+import java.util.concurrent.TimeUnit
+import javax.inject.Named
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NetworkModule {
+
+    private const val TAG_REQUEST = "RequestClient"
+    private const val READ_TIME_OUT: Long = 60
+    private const val CONNECTION_TIMEOUT: Long = 60
+
+    @Singleton
+    @Provides
+    fun provideLoggingInterceptor(): HttpLoggingInterceptor {
+        return HttpLoggingInterceptor(object : HttpLoggingInterceptor.Logger {
+            override fun log(message: String) {
+                Timber.tag(TAG_REQUEST).d(message)
+            }
+        }).apply {
+            level =
+                if (BuildConfig.DEBUG) HttpLoggingInterceptor.Level.BODY
+                else HttpLoggingInterceptor.Level.NONE
+        }
+    }
+
+    @Singleton
+    @Provides
+    fun provideOkHttpClientBuilder(interceptor: HttpLoggingInterceptor): OkHttpClient.Builder {
+        return OkHttpClient.Builder()
+            .addInterceptor(interceptor)
+            .addInterceptor(CommonHeadersInterceptor())
+            .readTimeout(READ_TIME_OUT, TimeUnit.SECONDS)
+            .connectTimeout(CONNECTION_TIMEOUT, TimeUnit.SECONDS)
+    }
+
+    @Singleton
+    @Provides
+    @Named("retrofit_star_wars_api")
+    fun provideRetrofitClient(clientBuilder: OkHttpClient.Builder): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BuildConfig.BASE_URL)
+            .client(clientBuilder.build())
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+}
