+++ app/src/main/java/com/example/luistovar/archapp/framework/network/CommonHeadersInterceptor.kt
@@ -1,0 +1,17 @@
+package com.example.luistovar.archapp.framework.network
+
+import okhttp3.Interceptor
+import okhttp3.Response
+
+class CommonHeadersInterceptor : Interceptor {
+
+    override fun intercept(chain: Interceptor.Chain): Response {
+        val request = chain.request()
+        val newRequest = request.newBuilder()
+            .addHeader("common_header", "value of common header")
+            .addHeader("common_header_wo", "value of common header two")
+            .build()
+        return chain.proceed(newRequest)
+
+    }
+}
