+++ app/src/main/java/com/example/luistovar/archapp/framework/presentation/MainActivity.kt
@@ -1,0 +1,16 @@
+package com.example.luistovar.archapp.framework.presentation
+
+import com.example.luistovar.archapp.R
+import com.example.luistovar.archapp.framework.presentation.common.basecomponents.BaseActivity
+import dagger.hilt.android.AndroidEntryPoint
+
+/**
+ * MainActivity
+ *
+ * Activity that contains the root view in which will be set all the fragments
+ */
+@AndroidEntryPoint
+class MainActivity : BaseActivity(R.layout.main_activity) {
+
+
+}
