+++ app/src/main/java/com/example/luistovar/archapp/framework/presentation/main/MainViewModel.kt
@@ -1,0 +1,10 @@
+package com.example.luistovar.archapp.framework.presentation.main
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class MainViewModel @Inject constructor(): ViewModel() {
+    // TODO: Implement the ViewModel
+}
