--- app/src/main/java/com/example/luistovar/archapp/presentation/MainActivity.kt
@@ -1,14 +1,0 @@
-package com.example.luistovar.archapp.presentation
-
-import com.example.luistovar.archapp.R
-import com.example.luistovar.archapp.presentation.common.basecomponents.BaseActivity
-
-/**
- * MainActivity
- *
- * Activity that contains the root view in which will be set all the fragments
- */
-class MainActivity : BaseActivity(R.layout.main_activity) {
-
-
-}
