--- app/src/main/java/com/example/luistovar/archapp/presentation/common/basecomponents/BaseActivity.kt
@@ -1,15 +1,0 @@
-package com.example.luistovar.archapp.presentation.common.basecomponents
-
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-
-/**
- * BaseActivity
- *
- * Base activity that contains commons methods and configuration that need or could be used
- * for all activities in the app
- */
-abstract class BaseActivity(@LayoutRes contentLayoutId: Int) : AppCompatActivity(contentLayoutId) {
-
-
-}
