--- app/src/main/java/com/example/luistovar/archapp/presentation/home/MainViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.luistovar.archapp.presentation.home
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
