+++ app/src/main/java/com/example/luistovar/archapp/usecases/PeopleListSwUseCase.kt
@@ -1,0 +1,10 @@
+package com.example.luistovar.archapp.usecases
+
+import com.example.luistovar.archapp.domain.models.PeopleListSw
+import com.example.luistovar.archapp.domain.models.Resource
+
+interface PeopleListSwUseCase {
+
+    suspend fun getPeopleListSwResource(): Resource<PeopleListSw?>
+
+}
