+++ app/src/main/res/layout/list_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.listdata.ListFragment">
+    tools:context=".framework.presentation.listdata.ListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rVPeople"
