+++ app/src/main/res/layout/main_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:viewBindingIgnore="true"
-    tools:context=".presentation.MainActivity">
+    tools:context=".framework.presentation.MainActivity">
 
 
     <androidx.fragment.app.FragmentContainerView
