+++ build.gradle
@@ -2,16 +2,18 @@
 buildscript {
     ext.kotlin_version = "1.5.10"
     ext.nav_version = "2.3.5"
-    ext.koin_version = "3.1.1"
+    ext.hilt_version = "2.37"
     repositories {
         google()
         mavenCentral()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:4.2.1"
+        classpath "com.android.tools.build:gradle:4.2.2"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         //Safe Args
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
+        //DI Hilt
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.28-alpha'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
